moved common sign function
