+++ libavfilter/vf_signature_cuda.c
@@ -19,15 +19,7 @@
 #include "scale_eval.h"
 #include "video.h"
 #include "signature.h"
-
-
-static const enum AVPixelFormat supported_formats[] = {
-    AV_PIX_FMT_YUV420P,
-    AV_PIX_FMT_NV12,
-    AV_PIX_FMT_YUV444P,
-    AV_PIX_FMT_YUV444P16,
-    AV_PIX_FMT_YUV422P,
-};
+#include "signature_lookup.c"
 
 #define DIV_UP(a, b) ( ((a) + (b) - 1) / (b) )
 #define BLOCKX 32
@@ -36,14 +28,13 @@ static const enum AVPixelFormat supported_formats[] = {
 #define H_SIGN 32
 #define PIXELS_SIGN 1024
 #define BLOCK_LCM (int64_t) 476985600
-
 #define CHECK_CU(x) FF_CUDA_CHECK_DL(ctx, s->hwctx->internal->cuda_dl, x)
 
 typedef struct CUDASignContext {
     const AVClass *class;
 
     char *filename;
-    int format;
+    int  format;
 
     StreamContext* streamcontexts;
 
@@ -59,6 +50,26 @@ typedef struct CUDASignContext {
 
     float param;
 } CUDASignContext;
+
+#define OFFSET(x) offsetof(CUDASignContext, x)
+#define FLAGS (AV_OPT_FLAG_FILTERING_PARAM|AV_OPT_FLAG_VIDEO_PARAM)
+static const AVOption options[] = {
+    { "filename",   "filename for output file",
+        OFFSET(filename), AV_OPT_TYPE_STRING, {.str = ""}, 0, NB_FORMATS-1, FLAGS },
+    { "format",     "set output format",
+        OFFSET(format),       AV_OPT_TYPE_INT, {.i64 = FORMAT_BINARY}, 0, 1, FLAGS , "format" },
+        { "binary", 0, 0, AV_OPT_TYPE_CONST, {.i64=FORMAT_BINARY}, 0, 0, FLAGS, "format" },
+        { "xml",    0, 0, AV_OPT_TYPE_CONST, {.i64=FORMAT_XML},    0, 0, FLAGS, "format" },
+    { NULL },
+};
+
+static const enum AVPixelFormat supported_formats[] = {
+    AV_PIX_FMT_YUV420P,
+    AV_PIX_FMT_NV12,
+    AV_PIX_FMT_YUV444P,
+    AV_PIX_FMT_YUV444P16,
+    AV_PIX_FMT_YUV422P,
+};
 
 static int cudasign_query_formats(AVFilterContext *ctx)
 {
@@ -233,244 +244,6 @@ static int run_cudaresize(AVFilterContext *ctx, AVFrame *in)
     return ret;
 }
 
-static int get_block_size(const Block *b)
-{
-    return (b->to.y - b->up.y + 1) * (b->to.x - b->up.x + 1);
-}
-
-static uint64_t get_block_sum(StreamContext *sc, uint64_t intpic[32][32], const Block *b)
-{
-    uint64_t sum = 0;
-
-    int x0, y0, x1, y1;
-
-    x0 = b->up.x;
-    y0 = b->up.y;
-    x1 = b->to.x;
-    y1 = b->to.y;
-
-    if (x0-1 >= 0 && y0-1 >= 0) {
-        sum = intpic[y1][x1] + intpic[y0-1][x0-1] - intpic[y1][x0-1] - intpic[y0-1][x1];
-    } else if (x0-1 >= 0) {
-        sum = intpic[y1][x1] - intpic[y1][x0-1];
-    } else if (y0-1 >= 0) {
-        sum = intpic[y1][x1] - intpic[y0-1][x1];
-    } else {
-        sum = intpic[y1][x1];
-    }
-    return sum;
-}
-
-static int cmp(const void *x, const void *y)
-{
-    const uint64_t *a = x, *b = y;
-    return *a < *b ? -1 : ( *a > *b ? 1 : 0 );
-}
-
-static void set_bit(uint8_t* data, size_t pos)
-{
-    uint8_t mask = 1 << 7-(pos%8);
-    data[pos/8] |= mask;
-}
-
-static int xml_export(AVFilterContext *ctx, StreamContext *sc, const char* filename)
-{
-    FineSignature* fs;
-    CoarseSignature* cs;
-    int i, j;
-    FILE* f;
-    unsigned int pot3[5] = { 3*3*3*3, 3*3*3, 3*3, 3, 1 };
-
-    f = fopen(filename, "w");
-    if (!f) {
-        int err = AVERROR(EINVAL);
-        char buf[128];
-        av_strerror(err, buf, sizeof(buf));
-        av_log(ctx, AV_LOG_ERROR, "cannot open xml file %s: %s\n", filename, buf);
-        return err;
-    }
-
-    /* header */
-    fprintf(f, "<?xml version='1.0' encoding='ASCII' ?>\n");
-    fprintf(f, "<Mpeg7 xmlns=\"urn:mpeg:mpeg7:schema:2001\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xsi:schemaLocation=\"urn:mpeg:mpeg7:schema:2001 schema/Mpeg7-2001.xsd\">\n");
-    fprintf(f, "  <DescriptionUnit xsi:type=\"DescriptorCollectionType\">\n");
-    fprintf(f, "    <Descriptor xsi:type=\"VideoSignatureType\">\n");
-    fprintf(f, "      <VideoSignatureRegion>\n");
-    fprintf(f, "        <VideoSignatureSpatialRegion>\n");
-    fprintf(f, "          <Pixel>0 0 </Pixel>\n");
-    fprintf(f, "          <Pixel>%d %d </Pixel>\n", sc->w - 1, sc->h - 1);
-    fprintf(f, "        </VideoSignatureSpatialRegion>\n");
-    fprintf(f, "        <StartFrameOfSpatialRegion>0</StartFrameOfSpatialRegion>\n");
-    /* hoping num is 1, other values are vague */
-    fprintf(f, "        <MediaTimeUnit>%d</MediaTimeUnit>\n", sc->time_base.den / sc->time_base.num);
-    fprintf(f, "        <MediaTimeOfSpatialRegion>\n");
-    fprintf(f, "          <StartMediaTimeOfSpatialRegion>0</StartMediaTimeOfSpatialRegion>\n");
-    fprintf(f, "          <EndMediaTimeOfSpatialRegion>%" PRIu64 "</EndMediaTimeOfSpatialRegion>\n", sc->coarseend->last->pts);
-    fprintf(f, "        </MediaTimeOfSpatialRegion>\n");
-
-    /* coarsesignatures */
-    for (cs = sc->coarsesiglist; cs; cs = cs->next) {
-        fprintf(f, "        <VSVideoSegment>\n");
-        fprintf(f, "          <StartFrameOfSegment>%" PRIu32 "</StartFrameOfSegment>\n", cs->first->index);
-        fprintf(f, "          <EndFrameOfSegment>%" PRIu32 "</EndFrameOfSegment>\n", cs->last->index);
-        fprintf(f, "          <MediaTimeOfSegment>\n");
-        fprintf(f, "            <StartMediaTimeOfSegment>%" PRIu64 "</StartMediaTimeOfSegment>\n", cs->first->pts);
-        fprintf(f, "            <EndMediaTimeOfSegment>%" PRIu64 "</EndMediaTimeOfSegment>\n", cs->last->pts);
-        fprintf(f, "          </MediaTimeOfSegment>\n");
-        for (i = 0; i < 5; i++) {
-            fprintf(f, "          <BagOfWords>");
-            for (j = 0; j < 31; j++) {
-                uint8_t n = cs->data[i][j];
-                if (j < 30) {
-                    fprintf(f, "%d  %d  %d  %d  %d  %d  %d  %d  ", (n & 0x80) >> 7,
-                                                                   (n & 0x40) >> 6,
-                                                                   (n & 0x20) >> 5,
-                                                                   (n & 0x10) >> 4,
-                                                                   (n & 0x08) >> 3,
-                                                                   (n & 0x04) >> 2,
-                                                                   (n & 0x02) >> 1,
-                                                                   (n & 0x01));
-                } else {
-                    /* print only 3 bit in last byte */
-                    fprintf(f, "%d  %d  %d ", (n & 0x80) >> 7,
-                                              (n & 0x40) >> 6,
-                                              (n & 0x20) >> 5);
-                }
-            }
-            fprintf(f, "</BagOfWords>\n");
-        }
-        fprintf(f, "        </VSVideoSegment>\n");
-    }
-
-    /* finesignatures */
-    for (fs = sc->finesiglist; fs; fs = fs->next) {
-        fprintf(f, "        <VideoFrame>\n");
-        fprintf(f, "          <MediaTimeOfFrame>%" PRIu64 "</MediaTimeOfFrame>\n", fs->pts);
-        /* confidence */
-        fprintf(f, "          <FrameConfidence>%d</FrameConfidence>\n", fs->confidence);
-        /* words */
-        fprintf(f, "          <Word>");
-        for (i = 0; i < 5; i++) {
-            fprintf(f, "%d ", fs->words[i]);
-            if (i < 4) {
-                fprintf(f, " ");
-            }
-        }
-        fprintf(f, "</Word>\n");
-        /* framesignature */
-        fprintf(f, "          <FrameSignature>");
-        for (i = 0; i< SIGELEM_SIZE/5; i++) {
-            if (i > 0) {
-                fprintf(f, " ");
-            }
-            fprintf(f, "%d ", fs->framesig[i] / pot3[0]);
-            for (j = 1; j < 5; j++)
-                fprintf(f, " %d ", fs->framesig[i] % pot3[j-1] / pot3[j] );
-        }
-        fprintf(f, "</FrameSignature>\n");
-        fprintf(f, "        </VideoFrame>\n");
-    }
-    fprintf(f, "      </VideoSignatureRegion>\n");
-    fprintf(f, "    </Descriptor>\n");
-    fprintf(f, "  </DescriptionUnit>\n");
-    fprintf(f, "</Mpeg7>\n");
-
-    fclose(f);
-    return 0;
-}
-
-static int binary_export(AVFilterContext *ctx, StreamContext *sc, const char* filename)
-{
-    FILE* f;
-    FineSignature* fs;
-    CoarseSignature* cs;
-    uint32_t numofsegments = (sc->lastindex + 44)/45;
-    int i, j;
-    PutBitContext buf;
-    /* buffer + header + coarsesignatures + finesignature */
-    int len = (512 + 6 * 32 + 3*16 + 2 +
-        numofsegments * (4*32 + 1 + 5*243) +
-        sc->lastindex * (2 + 32 + 6*8 + 608)) / 8;
-    uint8_t* buffer = av_malloc_array(len, sizeof(uint8_t));
-    if (!buffer)
-        return AVERROR(ENOMEM);
-
-    f = fopen(filename, "wb");
-    if (!f) {
-        int err = AVERROR(EINVAL);
-        char buf[128];
-        av_strerror(err, buf, sizeof(buf));
-        av_log(ctx, AV_LOG_ERROR, "cannot open file %s: %s\n", filename, buf);
-        av_freep(&buffer);
-        return err;
-    }
-    init_put_bits(&buf, buffer, len);
-
-    put_bits32(&buf, 1); /* NumOfSpatial Regions, only 1 supported */
-    put_bits(&buf, 1, 1); /* SpatialLocationFlag, always the whole image */
-    put_bits32(&buf, 0); /* PixelX,1 PixelY,1, 0,0 */
-    put_bits(&buf, 16, sc->w-1 & 0xFFFF); /* PixelX,2 */
-    put_bits(&buf, 16, sc->h-1 & 0xFFFF); /* PixelY,2 */
-    put_bits32(&buf, 0); /* StartFrameOfSpatialRegion */
-    put_bits32(&buf, sc->lastindex); /* NumOfFrames */
-    /* hoping num is 1, other values are vague */
-    /* den/num might be greater than 16 bit, so cutting it */
-    put_bits(&buf, 16, 0xFFFF & (sc->time_base.den / sc->time_base.num)); /* MediaTimeUnit */
-    put_bits(&buf, 1, 1); /* MediaTimeFlagOfSpatialRegion */
-    put_bits32(&buf, 0); /* StartMediaTimeOfSpatialRegion */
-    put_bits32(&buf, 0xFFFFFFFF & sc->coarseend->last->pts); /* EndMediaTimeOfSpatialRegion */
-    put_bits32(&buf, numofsegments); /* NumOfSegments */
-    /* coarsesignatures */
-    for (cs = sc->coarsesiglist; cs; cs = cs->next) {
-        put_bits32(&buf, cs->first->index); /* StartFrameOfSegment */
-        put_bits32(&buf, cs->last->index); /* EndFrameOfSegment */
-        put_bits(&buf, 1, 1); /* MediaTimeFlagOfSegment */
-        put_bits32(&buf, 0xFFFFFFFF & cs->first->pts); /* StartMediaTimeOfSegment */
-        put_bits32(&buf, 0xFFFFFFFF & cs->last->pts); /* EndMediaTimeOfSegment */
-        for (i = 0; i < 5; i++) {
-            /* put 243 bits ( = 7 * 32 + 19 = 8 * 28 + 19) into buffer */
-            for (j = 0; j < 30; j++) {
-                put_bits(&buf, 8, cs->data[i][j]);
-            }
-            put_bits(&buf, 3, cs->data[i][30] >> 5);
-        }
-    }
-    /* finesignatures */
-    put_bits(&buf, 1, 0); /* CompressionFlag, only 0 supported */
-    for (fs = sc->finesiglist; fs; fs = fs->next) {
-        put_bits(&buf, 1, 1); /* MediaTimeFlagOfFrame */
-        put_bits32(&buf, 0xFFFFFFFF & fs->pts); /* MediaTimeOfFrame */
-        put_bits(&buf, 8, fs->confidence); /* FrameConfidence */
-        for (i = 0; i < 5; i++) {
-            put_bits(&buf, 8, fs->words[i]); /* Words */
-        }
-        /* framesignature */
-        for (i = 0; i < SIGELEM_SIZE/5; i++) {
-            put_bits(&buf, 8, fs->framesig[i]);
-        }
-    }
-
-    flush_put_bits(&buf);
-    fwrite(buffer, 1, put_bits_count(&buf)/8, f);
-    fclose(f);
-    av_freep(&buffer);
-    return 0;
-}
-
-static int export(AVFilterContext *ctx, StreamContext *sc, int input)
-{
-    CUDASignContext* sic = ctx->priv;
-    char filename[1024];
-
-    if (av_strlcpy(filename, sic->filename, sizeof(filename)) >= sizeof(filename))
-        return AVERROR(EINVAL);
-
-    if (sic->format == FORMAT_XML) {
-        return xml_export(ctx, sc, filename);
-    } else {
-        return binary_export(ctx, sc, filename);
-    }
-}
 static int cudasign_filter_frame(AVFilterLink *link, AVFrame *in)
 {
     AVFilterContext         *ctx = link->dst;
@@ -480,31 +253,15 @@ static int cudasign_filter_frame(AVFilterLink *link, AVFrame *in)
     StreamContext           *sc = s->streamcontexts;
     FineSignature* fs;
 
-    static const uint8_t pot3[5] = { 3*3*3*3, 3*3*3, 3*3, 3, 1 };
-    /* indexes of words : 210,217,219,274,334  44,175,233,270,273  57,70,103,237,269  100,285,295,337,354  101,102,111,275,296
-    s2usw = sorted to unsorted wordvec: 44 is at index 5, 57 at index 10...
-    */
-    static const unsigned int wordvec[25] = {44,57,70,100,101,102,103,111,175,210,217,219,233,237,269,270,273,274,275,285,295,296,334,337,354};
-    static const uint8_t      s2usw[25]   = { 5,10,11, 15, 20, 21, 12, 22,  6,  0,  1,  2,  7, 13, 14,  8,  9,  3, 23, 16, 17, 24,  4, 18, 19};
-
-    uint8_t wordt2b[5] = { 0, 0, 0, 0, 0 }; /* word ternary to binary */
     uint64_t intpic[32][32];
     uint64_t rowcount;
-    uint64_t conflist[DIFFELEM_SIZE];
-    int f = 0, g = 0, w = 0;
+    int i, j, ret;
     int32_t dh1 = 1, dh2 = 1, dw1 = 1, dw2 = 1, a, b;
     int64_t denom;
-    int i, j, k, ternary;
-    uint64_t blocksum;
-    int blocksize;
-    int64_t th; /* threshold */
-    int64_t sum;
     int64_t *pp;
-
     int64_t precfactor;
     
     CUcontext dummy;
-    int ret = 0;
 
     ret = CHECK_CU(cu->cuCtxPushCurrent(s->hwctx->cuda_ctx));
     if (ret < 0)
@@ -577,115 +334,30 @@ static int cudasign_filter_frame(AVFilterLink *link, AVFrame *in)
     }
 
     denom = (sc->divide) ? 1 : dh1 * dh2 * dw1 * dw2;
-
-    for (i = 0; i < ELEMENT_COUNT; i++) {
-        const ElemCat* elemcat = elements[i];
-        int64_t* elemsignature;
-        uint64_t* sortsignature;
-
-        elemsignature = av_malloc_array(elemcat->elem_count, sizeof(int64_t));
-        if (!elemsignature)
-            return AVERROR(ENOMEM);
-        sortsignature = av_malloc_array(elemcat->elem_count, sizeof(int64_t));
-        if (!sortsignature) {
-            av_freep(&elemsignature);
-            return AVERROR(ENOMEM);
-        }
-
-        for (j = 0; j < elemcat->elem_count; j++) {
-            blocksum = 0;
-            blocksize = 0;
-            for (k = 0; k < elemcat->left_count; k++) {
-                blocksum += get_block_sum(sc, intpic, &elemcat->blocks[j*elemcat->block_count+k]);
-                blocksize += get_block_size(&elemcat->blocks[j*elemcat->block_count+k]);
-            }
-            sum = blocksum / blocksize;
-            if (elemcat->av_elem) {
-                sum -= 128 * precfactor * denom;
-            } else {
-                blocksum = 0;
-                blocksize = 0;
-                for (; k < elemcat->block_count; k++) {
-                    blocksum += get_block_sum(sc, intpic, &elemcat->blocks[j*elemcat->block_count+k]);
-                    blocksize += get_block_size(&elemcat->blocks[j*elemcat->block_count+k]);
-                }
-                sum -= blocksum / blocksize;
-                conflist[g++] = FFABS(sum * 8 / (precfactor * denom));
-            }
-
-            elemsignature[j] = sum;
-            sortsignature[j] = FFABS(sum);
-        }
-
-        /* get threshold */
-        qsort(sortsignature, elemcat->elem_count, sizeof(uint64_t), cmp);
-        th = sortsignature[(int) (elemcat->elem_count*0.333)];
-
-        /* ternarize */
-        for (j = 0; j < elemcat->elem_count; j++) {
-            if (elemsignature[j] < -th) {
-                ternary = 0;
-            } else if (elemsignature[j] <= th) {
-                ternary = 1;
-            } else {
-                ternary = 2;
-            }
-            fs->framesig[f/5] += ternary * pot3[f%5];
-
-            if (f == wordvec[w]) {
-                fs->words[s2usw[w]/5] += ternary * pot3[wordt2b[s2usw[w]/5]++];
-                if (w < 24)
-                    w++;
-            }
-            f++;
-        }
-        av_freep(&elemsignature);
-        av_freep(&sortsignature);
-    }
-
-    /* confidence */
-    qsort(conflist, DIFFELEM_SIZE, sizeof(uint64_t), cmp);
-    fs->confidence = FFMIN(conflist[DIFFELEM_SIZE/2], 255);
-
-    /* coarsesignature */
-    if (sc->coarsecount == 0) {
-        if (sc->curcoarsesig2) {
-            sc->curcoarsesig1 = av_mallocz(sizeof(CoarseSignature));
-            if (!sc->curcoarsesig1)
-                return AVERROR(ENOMEM);
-            sc->curcoarsesig1->first = fs;
-            sc->curcoarsesig2->next = sc->curcoarsesig1;
-            sc->coarseend = sc->curcoarsesig1;
-        }
-    }
-    if (sc->coarsecount == 45) {
-        sc->midcoarse = 1;
-        sc->curcoarsesig2 = av_mallocz(sizeof(CoarseSignature));
-        if (!sc->curcoarsesig2)
-            return AVERROR(ENOMEM);
-        sc->curcoarsesig2->first = fs;
-        sc->curcoarsesig1->next = sc->curcoarsesig2;
-        sc->coarseend = sc->curcoarsesig2;
-    }
-    for (i = 0; i < 5; i++) {
-        set_bit(sc->curcoarsesig1->data[i], fs->words[i]);
-    }
-    /* assuming the actual frame is the last */
-    sc->curcoarsesig1->last = fs;
-    if (sc->midcoarse) {
-        for (i = 0; i < 5; i++) {
-            set_bit(sc->curcoarsesig2->data[i], fs->words[i]);
-        }
-        sc->curcoarsesig2->last = fs;
-    }
-
-    sc->coarsecount = (sc->coarsecount+1)%90;
+    ret = calc_signature(ctx, sc, fs, intpic, denom, precfactor);
+    if (ret < 0)
+        goto fail;
 
     //passthrough
     return ff_filter_frame(outlink, in);
 fail:
     av_frame_free(&in);    
     return ret;
+}
+
+static int export(AVFilterContext *ctx, StreamContext *sc, int input)
+{
+    CUDASignContext* sic = ctx->priv;
+    char filename[1024];
+
+    if (av_strlcpy(filename, sic->filename, sizeof(filename)) >= sizeof(filename))
+        return AVERROR(EINVAL);
+
+    if (sic->format == FORMAT_XML) {
+        return xml_export(ctx, sc, filename);
+    } else {
+        return binary_export(ctx, sc, filename);
+    }
 }
 
 static av_cold int cudasign_init(AVFilterContext *ctx)
@@ -771,18 +443,6 @@ static av_cold void cudasign_uninit(AVFilterContext *ctx)
         av_freep(&s->streamcontexts);
     }    
 }
-
-#define OFFSET(x) offsetof(CUDASignContext, x)
-#define FLAGS (AV_OPT_FLAG_FILTERING_PARAM|AV_OPT_FLAG_VIDEO_PARAM)
-static const AVOption options[] = {
-    { "filename",   "filename for output file",
-        OFFSET(filename), AV_OPT_TYPE_STRING, {.str = ""}, 0, NB_FORMATS-1, FLAGS },
-    { "format",     "set output format",
-        OFFSET(format),       AV_OPT_TYPE_INT, {.i64 = FORMAT_BINARY}, 0, 1, FLAGS , "format" },
-        { "binary", 0, 0, AV_OPT_TYPE_CONST, {.i64=FORMAT_BINARY}, 0, 0, FLAGS, "format" },
-        { "xml",    0, 0, AV_OPT_TYPE_CONST, {.i64=FORMAT_XML},    0, 0, FLAGS, "format" },
-    { NULL },
-};
 
 static const AVClass cudasign_class = {
     .class_name = "cudasign",
